Move to rand.go
