+++ rand_1.10_test.go
@@ -1,0 +1,10 @@
+// Copyright 2018 Keybase, Inc. All rights reserved. Use of
+// this source code is governed by the included BSD license.
+
+package saltpack
+
+import "testing"
+
+func TestCryptoRandUint32(t *testing.T) {
+
+}
